// Frontend/src/Types/RoomResultPage.d.ts
export interface RoomCardType {
    roomTypeName:            string;
    maxCapacity:             string;
    doubleBeds:              number;
    singleBeds:              number;
    area:                    number;
    roomTypeId:              number;
    averageRoomRate:         number;
    carouselImages:          string[];
    rating:                  number;
    review:                  number;
    bestPromotion ?: string;
}

export interface RoomAPIResponse{
    roomCardDTOList : RoomCardType[],
    length : number
} 


export interface Promotion {
    promotion_id:          number;
    promotion_title:       string;
    promotion_description: string;
    is_deactivated:        boolean;
    minimum_days_of_stay:  number;
    price_factor:          number;
}




// Generated by https://quicktype.io

export interface Feedback {
    feedbackId:      string;
    name:            string;
    feedbackContent: string;
    rating:          number;
    roomTypeId:      number;
    helpful:         number;
    notHelpful:      number;
    createdAt:       Date;
    submitted:       boolean;
}